pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "AutoMorph"
version = "0.1.0"
description = "AIBILI fork of AutoMorph"
readme = "README.md"
requires-python = ">=3.8"
license = "MIT"
authors = [
    {name = "Gonçalo Ferreira", email = "gaferreira@aibili.pt"}
]
dependencies = [
    "numpy>=1.23.5",
    "opencv-python-headless==4.10.0.84",
    "pandas",
    "Pillow",
    "pyarrow>=19.0.1",

    "tqdm>=4.66.4",
    "scipy>=1.14.0",

    "scikit-image>=0.24.0",
    "fsspec>=2025.3.0",

    "h5py>=3.11.0",
    "matplotlib>=3.9.0"

]

[tool.setuptools]
packages = []